Clean up warnings
Got rid of unused variables and added new to variables that hid existing members. @ImWesMan take a look at toggleSizeChoice in Start() you make a new variable called enabled instead of setting the existing one. I removed it but feel free to update it so that the code looks cleaner.